+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 # 6.4.9
 * Add feature to save logs to sdcard for debugging.
 * Fix bug causing long running Farmwares to fail.
+* Make sure to clear eeprom before flashing Arduino firmware.
 * Update Linux system layer.
 
 # 6.4.8